refactor: Rename `lib/logsCollection.js`
