+++ lib/logs-collection.js
