+++ lib/logs-collection.test.js
@@ -12,7 +12,7 @@ describe('logsCollection', () => {
 
   before(() => {
     getLogDestinationStub = sinon.stub().resolves({ destinationArn: 'arn:logdest' });
-    logsCollection = proxyquire('./logsCollection', {
+    logsCollection = proxyquire('./logs-collection', {
       './client-utils': {
         getPlatformClientWithAccessKey: () => ({
           logDestinations: {
