+++ lib/plugin.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const { ServerlessSDK } = require('@serverless/platform-client');
 const log = require('./log');
 const errorHandler = require('./error-handler');
-const logsCollection = require('./logsCollection');
+const logsCollection = require('./logs-collection');
 const login = require('./login');
 const logout = require('./logout');
 const wrap = require('./wrap');
