+++ lib/plugin.test.js
@@ -43,7 +43,7 @@ describe('plugin', () => {
     setApiGatewayAccessLogFormat = sinon.spy();
 
     ServerlessEnterprisePlugin = proxyquire('./plugin', {
-      './logsCollection': logsCollection,
+      './logs-collection': logsCollection,
       './wrap': wrap,
       './wrapClean': wrapClean,
       './removeDestination': removeDestination,
